Fixed some directory issues, and fixed gerber files.
Removed the Project Outputs folder and rearranged it's files to match standards. Removed the extra .pdf files made for presenting, and removed the History and _Previews directories (They will be back in the next commit). Also changed .1uF values to 0.1uF in Altium. Finally, gerber files are now working, although there are likely an excess of unnecessary gerber files.