Added README files.
The README file located in "AMDC-Hardware\Accessories\ExpansionBoard_Knob\REV20220427A". This README is likely complete, however some of the links may need to be updated. The image links are also set to work once the files are merged into the develop branch, but don't currently work.

The second README is located in "GitHub\AMDC-Hardware\Accessories\ExpansionBoard_Knob\REV20220427A". This README has a couple of unfilled inputs that I need to update later.